+++ gameserver/account/CMakeLists.txt
@@ -5,11 +5,13 @@ project(account)
 add_library(account
   "export/account.h"
   "src/account.cc"
+  "src/account_loader.cc"
+  "src/account_loader.h"
 )
 
 target_include_directories(account PUBLIC "export")
 
 target_link_libraries(account PRIVATE
-  io
   utils
+  rapidxml
 )
