+++ gameserver/utils/CMakeLists.txt
@@ -4,8 +4,11 @@ project(utils)
 
 set(FILES
   "export/config_parser.h"
+  "export/data_loader.h"
+  "export/file_reader.h"
   "export/logger.h"
   "export/tick.h"
+  "src/data_loader.cc"
   "src/logger.cc"
   "src/tick.cc"
 )
@@ -16,4 +19,10 @@ if (CMAKE_C_COMPILER MATCHES ".*emcc")
 endif()
 
 add_library(utils ${FILES})
+
+target_link_libraries(utils PRIVATE
+  common
+  rapidxml
+)
+
 target_include_directories(utils PUBLIC "export")
