+++ gameserver/wsclient/CMakeLists.txt
@@ -7,6 +7,8 @@ add_executable(wsclient
   "src/graphics.h"
   "src/network.cc"
   "src/network.h"
+  "src/sprite_loader.cc"
+  "src/sprite_loader.h"
   "src/texture.cc"
   "src/texture.h"
   "src/types.h"
@@ -17,10 +19,8 @@ add_executable(wsclient
 
 target_link_libraries(wsclient PRIVATE
   common
-  gameengine
-  io
-  network
-  protocol
+  network_packet
+  protocol_client
   utils
 )
 
@@ -30,13 +30,14 @@ target_link_options(wsclient PRIVATE --preload-file "../files/")
 
 add_executable(itemview
   "src/itemview.cc"
+  "src/sprite_loader.cc"
+  "src/sprite_loader.h"
   "src/texture.cc"
   "src/texture.h"
 )
 
 target_link_libraries(itemview PRIVATE
   common
-  io
   utils
 )
