+++ gameserver/wsclient/src/sprite_loader.h
